fix #7034 mapping for Java enums
